backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.2
plotly==5.23.0
pytest==8.3.2
httpx==0.27.2
pytest-asyncio==0.23.8
ollama>=0.4
google-auth>=2.0
google-auth-oauthlib>=1.0
google-api-python-client>=2.0
